create `ObjectWithPaths`, remove other custom classes & exceptions
