+++ langworld_db_data/removers/listed_value_remover.py
@@ -1,13 +1,13 @@
+from langworld_db_data import ObjectWithPaths
 from langworld_db_data.constants.literals import ID_SEPARATOR
 from langworld_db_data.filetools.csv_xls import read_dicts_from_csv, write_csv
-from langworld_db_data.removers.remover import Remover, RemoverError
 
 
-class ListedValueRemoverError(RemoverError):
+class ListedValueRemoverError(Exception):
     pass
 
 
-class ListedValueRemover(Remover):
+class ListedValueRemover(ObjectWithPaths):
     def remove_listed_value(
         self,
         id_of_value_to_remove: str,
